Update to isLessThan
Currently isLessThan can return the wrong value if the exponent differs by one.

e.g. 9*10^3 > 0.1*10^4, but the code currently thinks that it's smaller, as the exponent is smaller. I've added a check when exponents differ by one.

The problem might persist with other exponents, but those should be fixable with normalization.